Use m_webView->size() for viewport update https://bugs.webkit.org/show_bug.cgi?id=102764
Patch by Tien-Ren Chen <trchen@chromium.org> on 2012-11-21
Reviewed by Adam Barth.

m_webView->client()->windowRect() returns the outer rect of the browser
window, which includes the decorations such as the title bar.
We should use the size of the WebView for layout and rendering purpose.

Existing tests are updated to work without windowRect().

* src/ChromeClientImpl.cpp:
(WebKit::ChromeClientImpl::dispatchViewportPropertiesDidChange):
* tests/WebFrameTest.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@135442 268f45cc-cd09-0410-ab3c-d52691b4dbfc