[Qt] REGRESSION(r135515): It made 3-3 tests fail and crash https://bugs.webkit.org/show_bug.cgi?id=103073
Rubber-stamped by Simon Hausmann.

Source/WebKit/qt:

Removing the QNetworkAccessManager member in FrameNetworkingContextQt
was right, but we need to be careful as to how we fetch it back.
Since the WebCore Page is not always valid, get the QNAM instance
straight from the frame adapter through its page adapter instead.

* WebCoreSupport/ChromeClientQt.cpp: [Style nit] Remove stray emit.
* WebCoreSupport/FrameNetworkingContextQt.cpp:
(WebCore::FrameNetworkingContextQt::networkAccessManager):

LayoutTests:

Unskip the tests that had regressed.

* platform/qt/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@135532 268f45cc-cd09-0410-ab3c-d52691b4dbfc