Caret is painted horizontally in vertical writing mode when there are no visible text https://bugs.webkit.org/show_bug.cgi?id=102359
Patch by Arpita Bahuguna <arpitabahuguna@gmail.com> on 2012-11-30
Reviewed by Ryosuke Niwa.

Source/WebCore:

In the vertical writing mode the caret rect on an empty editable box or
an empty line is rendered as per the horizontal writing mode.

This changes though when any text is added; the caret is then displayed
in accordance with the vertical writing mode.

Test: editing/selection/caret-alignment-for-vertical-text.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::localCaretRect):
* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::localCaretRectForEmptyElement):
The caret rect should be transposed for the vertical writing mode. Its
width and height values need to be interchanged for it to be oriented
as per the vertical text.

LayoutTests:

* editing/selection/caret-alignment-for-vertical-text-expected.txt: Added.
* editing/selection/caret-alignment-for-vertical-text.html: Added.
Test to verify that the orientation of the caret on an empty vertical text
line or container is same as that for a non-empty container.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136225 268f45cc-cd09-0410-ab3c-d52691b4dbfc