[EFL][Qt][WK2] Fixed position elements are not always fixed https://bugs.webkit.org/show_bug.cgi?id=103452
Reviewed by Simon Fraser.

The code figuring out whether fixed position layers are inside
the visible viewport, assumes that the visible viewport is always
the size of the layout viewport. This assumption doesn't hold with
how the Qt and EFL tiled backing store and coordinated graphics
works, so instead using the visibleContentsRect(), which provides
the right values in all cases.

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForPosition):

    Use visibleContentsRect instead of scrollOffsetForFixedPosition().

    scrollOffsetForFixedPosition() is needed sometimes in the Mac code
    because visibleContentRect() will return negative offsets when you
    are in the rubber-band phase of a scroll on the Mac.

    However that is not an issue here.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136452 268f45cc-cd09-0410-ab3c-d52691b4dbfc