[Qt][Mac] Fix libxslt and libxml2 config tests https://bugs.webkit.org/show_bug.cgi?id=104164
Reviewed by Simon Hausmann.

Source/WebCore:

Since libxml2 is a dependency for libxslt and is not used
standalone the configurations for it should also depend on
whether XSLT is enabled.
Also avoid using pkg-config on Mac, instead use direct
include paths and add needed libraries to the linker.

No new tests needed.

* WebCore.pri:

Tools:

We should not use pkg-config on Mac instead use direct include
paths and add needed libraries to the linker options to detect
libxslt and libxml2 provided by the system.
Previously we would always fall back to qtxmlpatterns.

* qmake/config.tests/libxml2/libxml2.pro:
* qmake/config.tests/libxslt/libxslt.pro:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136837 268f45cc-cd09-0410-ab3c-d52691b4dbfc