Strange results calculating a square root in a loop https://bugs.webkit.org/show_bug.cgi?id=104247 <rdar://problem/12826880>
Reviewed by Oliver Hunt.

Source/JavaScriptCore: 

Fixed the CFG simplification phase to ignore dead GetLocals in the first of the blocks
under the merge. This fixes the assertion, and is also cleaner: our general rule is
to not "revive" things that we've already proved to be dead.
        
Also fixed some rotted debug code.

* dfg/DFGCFGSimplificationPhase.cpp:
(JSC::DFG::CFGSimplificationPhase::fixPossibleGetLocal):
* dfg/DFGStructureCheckHoistingPhase.cpp:
(JSC::DFG::StructureCheckHoistingPhase::run):

LayoutTests: 

* fast/js/dfg-cfg-simplify-redundant-dead-get-local-expected.txt: Added.
* fast/js/dfg-cfg-simplify-redundant-dead-get-local.html: Added.
* fast/js/jsc-test-list:
* fast/js/script-tests/dfg-cfg-simplify-redundant-dead-get-local.js: Added.
(getDist):
(calcError):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136989 268f45cc-cd09-0410-ab3c-d52691b4dbfc