Fix webgl-depth-texture.html to match the test expectation. https://bugs.webkit.org/show_bug.cgi?id=104259.
Patch by Kondapally Kalyan <kalyan.kondapally@intel.com> on 2012-12-07
Reviewed by Kenneth Russell.

The text expectation of this testcase expects to check for FrameBuffer status and compare the value
using ShouldBe after creating the framebuffer.
Currently, we do it only in case of adding a color buffer. This patch ensures that the Framebuffer
Status check using ShouldBe is done in all cases.

* fast/canvas/webgl/webgl-depth-texture.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@136990 268f45cc-cd09-0410-ab3c-d52691b4dbfc