[EFL] Unit tests process hanging on WK2 Release bots https://bugs.webkit.org/show_bug.cgi?id=105021
Reviewed by Kenneth Rohde Christiansen.

Do not allocate memory in the middle of a fork()/exec().
EFL uses global FastMalloc new operator and when we fork(), the
process is in a very delicate state. We were being lucky most of the
time, but sometimes FastMalloc enters in a busy wait and hangs
because it's data structures cannot be trusted after the fork().

This explains why sometimes we see UIProcess hanging on the bots:
they are in fact UIProcess trying to exec() into a WebProcess but
hanged just after the fork(). CMake test runner kills the original
UIProcess due to timeout and the buggy one stays forever.

* PlatformEfl.cmake:
Set the timeout back to the original value.

* UIProcess/Launcher/efl/ProcessLauncherEfl.cpp:
(WebKit::ProcessLauncher::launchProcess):
I'm keeping the original behavior, including the call to
system() when wrapping the WebProcess on Debug mode. This
should be changed to exec() in another patch.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137890 268f45cc-cd09-0410-ab3c-d52691b4dbfc