GraphicsLayer's repaint count should update even when debugging option to show it is off https://bugs.webkit.org/show_bug.cgi?id=105178
Patch by Antoine Quint <graouts@apple.com> on 2012-12-17
Reviewed by Simon Fraser.

Update the repain count even if the debug option to show them is turned off.
This allows the count to be accurate in between toggles of the debug option
and to allow the InspectorLayerTreeAgent to get accurate results any time
the layer tree is requested (see webkit.org/b/105024).

* platform/graphics/mac/WebLayer.mm:
(drawLayerContents):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137915 268f45cc-cd09-0410-ab3c-d52691b4dbfc