Don't allow edge TileCache tiles to be larger than necessary https://bugs.webkit.org/show_bug.cgi?id=105195
Reviewed by Anders Carlsson.

The TileCache would allow edge tiles to be larger than the required size, with the assumption
that the extra space would never be visible. However, for content tiled layers, it can be,
and it doesn't get correctly cleared. So always keep the edge tiles at the specified size.

* platform/graphics/ca/mac/TileCache.mm:
(WebCore::TileCache::revalidateTiles):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137923 268f45cc-cd09-0410-ab3c-d52691b4dbfc