Butterfly::growArrayRight shouldn't be called on null Butterfly objects https://bugs.webkit.org/show_bug.cgi?id=105221
Reviewed by Filip Pizlo.

Currently we depend upon the fact that Butterfly::growArrayRight works with null Butterfly 
objects purely by coincidence. We should add a new static function that null checks the old 
Butterfly object and creates a new one if it's null, or calls growArrayRight if it isn't for 
use in the couple of places in JSObject that expect such behavior to work.

* runtime/Butterfly.h:
(Butterfly):
* runtime/ButterflyInlines.h:
(JSC::Butterfly::createOrGrowArrayRight):
(JSC):
* runtime/JSObject.cpp:
(JSC::JSObject::createInitialIndexedStorage):
(JSC::JSObject::createArrayStorage):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137961 268f45cc-cd09-0410-ab3c-d52691b4dbfc