[BlackBerry] Use midpoint for fixed position heuristic https://bugs.webkit.org/show_bug.cgi?id=105329
Patch by Andrew Lo <anlo@rim.com> on 2012-12-18
Reviewed by Rob Buis.
Internally reviewed by Joe Mason.

Use the midpoint of the layer when determining whether the
fixed-position layer is on the top or bottom half of the screen
rather than the top of the layer.
Internal PR 237073.

No new tests are needed.

* platform/graphics/blackberry/LayerRenderer.cpp:
(WebCore::LayerRenderer::updateLayersRecursive):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138056 268f45cc-cd09-0410-ab3c-d52691b4dbfc