Web Inspector: AuditLauncherView UI components should be disabled until ongoing audit finishes https://bugs.webkit.org/show_bug.cgi?id=105389
Reviewed by Alexander Pavlov.

While the Audit request is on-going, UI components on the AuditLauncherView are disabled
until the request is completed or explicitly stopped. Also changed the _categoriesElement
from div to fieldset and modified the associated CSS file accordingly.

No new tests as this is a UI change.

* inspector/front-end/AuditLauncherView.js:
(WebInspector.AuditLauncherView.prototype._setAuditRunning):
(WebInspector.AuditLauncherView.prototype._toggleUIComponents):
* inspector/front-end/auditsPanel.css:
(.audit-launcher-view fieldset.audit-categories-container):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138122 268f45cc-cd09-0410-ab3c-d52691b4dbfc