[Shadow] TITLE elements in Shadow DOM should not affect document.title attribute https://bugs.webkit.org/show_bug.cgi?id=85864
Reviewed by Ryosuke Niwa.

Source/WebCore:

Modified HTMLTitleElement to check whether a title element is in a
shadow tree or not when it is inserted or removed. Now if a title
element has been just removed from or inserted into a shadow tree,
document.title is not updated independent of inDocument().  If the
title element is not in a shadow tree and in document,
document.title is updated.

Test: fast/dom/shadow/title-element-in-shadow.html

* html/HTMLTitleElement.cpp:
(WebCore::HTMLTitleElement::insertedInto):
Added a condition: isInShadowTree() to the code which checks
inDocument or not.
(WebCore::HTMLTitleElement::removedFrom):
Added a condition: insertionPoint.isInShadowTree() to the code which
checks an insertion point is in a document or not.
(WebCore::HTMLTitleElement::childrenChanged):
Added a condition: isInShadowTree() before setTitle.
If not isInShadowTree, modified to removeTitle from document.

LayoutTests:

* fast/dom/shadow/title-element-in-shadow-expected.txt: Added.
* fast/dom/shadow/title-element-in-shadow.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138130 268f45cc-cd09-0410-ab3c-d52691b4dbfc