::first-letter { overflow: -webkit-paged-y } causes crash https://bugs.webkit.org/show_bug.cgi?id=105393
Reviewed by Beth Dakin.

Source/WebCore:

Should check whether e is available or not before invoking
e->hasTagName in adjustRenderStyle in StyleResolver.cpp.

Test: fast/css/pseudo-element-opagedxy-crash.html

* css/StyleResolver.cpp:
(WebCore::StyleResolver::adjustRenderStyle):
If adjustRenderStyle is invoked in pseudoStyleForElement, the given
element is 0. So e is not available. However if the given style's
overflowY is OPAGEDX or OPAGEDY, adjustRenderStyle doesn't check
whether e is available or not before e->hasTagName.
This causes a crash.

LayoutTests:

* fast/css/pseudo-element-opagedxy-crash-expected.txt: Added.
* fast/css/pseudo-element-opagedxy-crash.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138451 268f45cc-cd09-0410-ab3c-d52691b4dbfc