Regression (r132935): WebKit breaks between <nobr> tags https://bugs.webkit.org/show_bug.cgi?id=11857
Reviewed by David Hyatt.

Source/WebCore:

The fix for bug 29648 in http://trac.webkit.org/changeset/132935 was too broad - it caused us
to find linebreaks in autowrap containers between non-wrapping elements where there was no
whitespace to allow the break. This reverts the code change in r132935 and ensures we attempt
to move below floats when inside an autowrap container but leaves the treatment of non-wrapping
elements in an autowrap container unaffected.

Test: fast/inline/break-between-nobr.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::nextSegmentBreak):

LayoutTests:

* fast/inline/break-between-nobr-expected.html: Added.
* fast/inline/break-between-nobr.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138630 268f45cc-cd09-0410-ab3c-d52691b4dbfc