CSP: Blocked 'track' sources should only log once to the console. https://bugs.webkit.org/show_bug.cgi?id=106091
Reviewed by Eric Carlson.

Source/WebCore:

HTMLTrackElement posts a message to the console when track data is
blocked by Content Security Policy. Now that we're logging inside the
ContentSecurityPolicy object itself, there's no need for the
duplication. This patch drops the extra log message (and, as a
side-effect, adds a test that ensures that 'track' content is governed
by the 'media-src' directive).

Test: http/tests/security/contentSecurityPolicy/media-src-track-block.html

* html/HTMLTrackElement.cpp:
(WebCore::HTMLTrackElement::canLoadUrl):

LayoutTests:

* http/tests/security/contentSecurityPolicy/media-src-track-block-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/media-src-track-block.html: Added.
* http/tests/security/contentSecurityPolicy/resources/track.srt: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138820 268f45cc-cd09-0410-ab3c-d52691b4dbfc