CSP 1.1: Experiment with adding line numbers to violation reports. https://bugs.webkit.org/show_bug.cgi?id=105374
Reviewed by Adam Barth.

Source/WebCore:

Line numbers and URLs were added to the CSP 1.1 spec in
https://dvcs.w3.org/hg/content-security-policy/rev/e0270baace3d, and
already exist (and have for quite some time) in Firefox's
implementation. Based on developer feedback, it seems reasonable to
experiment a bit with adding them to WebKit's violation reports.
This patch generates a stack trace when constructing the violation
report object (iff a violation report object ought to be generated)
and adds a line number and URL to the object if possible.

Tests: http/tests/security/contentSecurityPolicy/report-uri-from-inline-javascript.html
       http/tests/security/contentSecurityPolicy/report-uri-from-javascript.html

* page/ContentSecurityPolicy.cpp:
(WebCore::ContentSecurityPolicy::reportViolation):
    Generate a stack trace when building the violation report: if it
    has relevant detail, add the line number and URL to the report
    object.

LayoutTests:

* http/tests/security/contentSecurityPolicy/report-uri-from-inline-javascript-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/report-uri-from-inline-javascript.html: Added.
* http/tests/security/contentSecurityPolicy/report-uri-from-javascript-expected.txt: Added.
* http/tests/security/contentSecurityPolicy/report-uri-from-javascript.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138834 268f45cc-cd09-0410-ab3c-d52691b4dbfc