HTMLTreeBuilder should ASSERT that it is on the main thread before touching elements https://bugs.webkit.org/show_bug.cgi?id=106128
Reviewed by Eric Seidel.

Even if we move the HTML parser onto a background thread, we're likely
to keep fragment parsing on the main thread. These code paths touch
elements (which only exist on the main thread) but are only used during
fragment parsing. This patch adds ASSERTs to document that invariant.

* html/parser/HTMLTreeBuilder.cpp:
(WebCore::closestFormAncestor):
(WebCore::HTMLTreeBuilder::HTMLTreeBuilder):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138855 268f45cc-cd09-0410-ab3c-d52691b4dbfc