Clear pending container size requests as early as possible https://bugs.webkit.org/show_bug.cgi?id=106162
Reviewed by Eric Seidel.

http://trac.webkit.org/changeset/137981 contained a bug where pending container size
requests were only cleared for images that use container sizes. Bitmap images do not use
container sizes so these requests were not being cleared. This patch does the trivial
refactoring so that container size requests are cleared as early as possible.

No new tests as this is just a refactoring of some messy code added in r137981.

* loader/cache/CachedImage.cpp:
(WebCore::CachedImage::createImage):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138976 268f45cc-cd09-0410-ab3c-d52691b4dbfc