[chromium] Add a notification when a WebWidget's WebLayerTreeView is about to go away https://bugs.webkit.org/show_bug.cgi?id=106495
Patch by James Robinson <jamesr@chromium.org> on 2013-01-09
Reviewed by Adrienne Walker.

The WebWidget / WebLayerTreeView ownership is in a slightly awkward place. I'm moving the ownership
of the WebLayerTreeView to the WebWidgetClient, but to stage things sanely the WebLayerTreeView's
client is sill owned by the WebWidget implementation. Thus to shut down cleanly we need an explicit
notification to the WebWidget that the WebLayerTreeView is going to go away.

This API can go away once the WebWidget implementation does not provide the WebLayerTreeViewClient,
which will take a bit more refactoring.

* public/WebWidget.h:
(WebWidget):
(WebKit::WebWidget::willCloseLayerTreeView):
* src/WebViewImpl.cpp:
(WebKit::WebViewImpl::willCloseLayerTreeView):
(WebKit):
* src/WebViewImpl.h:
(WebViewImpl):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139252 268f45cc-cd09-0410-ab3c-d52691b4dbfc