[Chromium] Always enable autocomplete for password fields https://bugs.webkit.org/show_bug.cgi?id=104600
Patch by Yue Zhang <zysxqn@google.com> on 2013-01-09
Reviewed by Darin Fisher.

Don't check autocomplete in webkit code. Rather, we check it in chrome code in the following way: if the password field is chrome generated password, we ignore autocomplete=off and always fill the password; otherwise, we respect the autocomplete set. Since this is a chrome only feature, we make it configurable (default to false but enable this in chrome code).

* src/WebPasswordFormUtils.cpp:
(WebKit::findPasswordFormFields):
* src/WebPasswordFormUtils.h:
(WebKit):
(WebKit::findPasswordFormFields):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139253 268f45cc-cd09-0410-ab3c-d52691b4dbfc