HTMLMediaElement::resume() should schedule a load rather than load immediately https://bugs.webkit.org/show_bug.cgi?id=106587
Reviewed by Eric Carlson.

It is not always safe to start the jobs at the time resume() is called. That is why the jobs are
suspended. It can also be suspended again right after being resumed.

No new tests as there is no visible functional change, also it is a theoretically good-to-have change
that doesn't fix any known cross-platform issue.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::resume): Replace load() with scheduleLoad().


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139371 268f45cc-cd09-0410-ab3c-d52691b4dbfc