Fix a typo error in the comments in PlatformContextSkia.h https://bugs.webkit.org/show_bug.cgi?id=105612
Patch by Qiankun Miao <qiankun.miao@intel.com> on 2013-01-14
Reviewed by Stephen White.

"If false we're rendering to a GraphicsContext for a web page, if false
we're not (as is the case when rendering to a canvas object)." is
self-contradictory. The second "flase" in the comments should be
"true". If true, we're rendering to an ImageBuffer which has a canvas
object.

* platform/graphics/skia/PlatformContextSkia.h:
(PlatformContextSkia):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139591 268f45cc-cd09-0410-ab3c-d52691b4dbfc