REGRESSION (r137487): Crashes in editing/execCommand/indent-paragraphs.html on GTK, EFL https://bugs.webkit.org/show_bug.cgi?id=105042
Reviewed by Gustavo Noronha Silva.

Source/WebCore:

Explicitly handle the situation where the creation of a SoupURI fails. This
can happen if the URI is invalid. In that case the constructor returns null.

No new tests. This patch unskips a failing test.

* platform/network/soup/ResourceHandleSoup.cpp:
(WebCore::createSoupRequestAndMessageForHandle): Handle a null Soup URI.

LayoutTests:

Unskip a test which is now passing.

* platform/gtk/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140006 268f45cc-cd09-0410-ab3c-d52691b4dbfc