Teach Functional.h about WeakPtr https://bugs.webkit.org/show_bug.cgi?id=107105
Reviewed by Anders Carlsson.

A common pattern in cross-thread communication is to call member
functions of an object on a remote thread. If the caller's reference to
the object on the remote thread is a WeakPtr, the caller usually wants
to validate that the object still exists when the call actually takes
place.

It's possible to do this manually for every cross-thread call, but that
is tiresome and error prone. Instead, we can teach bind to validate
WeakPtr arguments when passed as the "this" parameter to a member
function.

* wtf/Functional.h:
(WTF::ParamStorageTraits::validate):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140040 268f45cc-cd09-0410-ab3c-d52691b4dbfc