[V8] Create a persistent wrapper for Window.prototype and innerGlobalObject https://bugs.webkit.org/show_bug.cgi?id=107253
Reviewed by Adam Barth.

This is one of steps to avoid hitting an ASSERT()
that will be added in https://bugs.webkit.org/show_bug.cgi?id=107137 .
We should have a persistent wrapper for all DOM objects
and set a class id on the wrapper.

No tests. No change in behavior.

* bindings/v8/V8DOMWindowShell.cpp:
(WebCore::V8DOMWindowShell::installDOMWindow):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140270 268f45cc-cd09-0410-ab3c-d52691b4dbfc