Refactoring: Use AtomicString for an InputType::create argument https://bugs.webkit.org/show_bug.cgi?id=107791
Reviewed by Kentaro Hara.

A string passed to InputType::create is a 'type' attribute value, which
is an AtomicString. Also, InputTypeFactoryMap is created with members of
InputTypeName, which are AtomicStrings. We had better use AtomicStrings
for them.

No new tests. This is just a refactoring.

* html/InputType.cpp:
Change the key type of InputTypeFactoryMap from String to AtomicString.
(WebCore::InputType::create):
Change an argument type from const String& to const AtomicString&.
* html/InputType.h:
(InputType): Ditto.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140665 268f45cc-cd09-0410-ab3c-d52691b4dbfc