[V8] Keep a RefPtr<SerializedScriptValue*> when we call serialize()/deserialize() (part 1) https://bugs.webkit.org/show_bug.cgi?id=107891
Reviewed by Abhishek Arya.

If you use a raw SerializedScriptValue* for serialize()/deserialize(),
it can potentially cause a use-after-free. This is because serialize()/
deserialize() can destruct a RefPtr of the SerializedScriptValue*,
depending on data that is serialized/deserialized. So we should keep a
RefPtr<SerializedScriptValue*> when we call serialize()/deserialize().
(See https://bugs.webkit.org/show_bug.cgi?id=107792 for more details.)

No tests. This is just a just-in-case fix. I couldn't find any bug
even in an ASAN build.

* bindings/v8/custom/V8CustomEventCustom.cpp:
(WebCore::V8CustomEvent::detailAccessorGetter):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140887 268f45cc-cd09-0410-ab3c-d52691b4dbfc