Added TriState to WTF and started using it in one place https://bugs.webkit.org/show_bug.cgi?id=108628
Reviewed by Beth Dakin.

../JavaScriptCore: 

* runtime/PrototypeMap.h:
(JSC::PrototypeMap::isPrototype): Use TriState instead of boolean. In
response to review feedback, this is an attempt to clarify that our
'true' condition is actually just a 'maybe'.

* runtime/PrototypeMap.h:
(PrototypeMap):
(JSC::PrototypeMap::isPrototype):

../WebCore: 

* editing/EditingStyle.h:
(WebCore): Moved TriState to WTF so it can be used in more places.

../WTF: 

Useful for expressing "maybe" conditions in a boolean context.

* WTF.xcodeproj/project.pbxproj:
* wtf/TriState.h: Added.
(WTF):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141588 268f45cc-cd09-0410-ab3c-d52691b4dbfc