class="cue" is getting some default style https://bugs.webkit.org/show_bug.cgi?id=108752
Source/WebCore:

Patch by Dima Gorbik <dgorbik@apple.com> on 2013-02-04
Reviewed by Dean Jackson.

The value variable inside the selector is used to store different information depending on the
type of the selector so we have to check explicitly that the selector we apply filtering to
matches a pseudo element.

Existing tests modified to cover this case.

* css/RuleSet.cpp:
(WebCore::determinePropertyWhitelistType):

LayoutTests:

Patch by Dima Gorbik <dgorbik@apple.com> on 2013-02-04
Reviewed by Dean Jackson.

* media/track/track-css-property-whitelist-expected.txt:
* media/track/track-css-property-whitelist.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141806 268f45cc-cd09-0410-ab3c-d52691b4dbfc