Source/WTF: Make StringBuilder::toAtomicString() consistent with StringBuilder::toString() for strings of length zero https://bugs.webkit.org/show_bug.cgi?id=108894
Patch by Benjamin Poulain <bpoulain@apple.com> on 2013-02-05
Reviewed by Andreas Kling.

* wtf/text/StringBuilder.h:
(WTF::StringBuilder::toAtomicString): The function was returning the nullAtom for strings of length zero.
This is inconsistent with StringBuilder::toString() which always return an empty string.

This patch unifies the behavior.

Tools: Make StringBuilder::toAtomicString() consistent with StringBuilder::toString() for strings of null length
https://bugs.webkit.org/show_bug.cgi?id=108894

Patch by Benjamin Poulain <bpoulain@apple.com> on 2013-02-05
Reviewed by Andreas Kling.

* TestWebKitAPI/Tests/WTF/StringBuilder.cpp:
Extend the tests to check toAtomicString() on an empty builder.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@141917 268f45cc-cd09-0410-ab3c-d52691b4dbfc