Unreviewed. More VS2010 WebKit solution touchups.
* win/tools/vsprops/common.props:
Move an ignored warning from a project to common properties.

* WebKit.vcxproj/WebKitExportGenerator/WebKitExportGenerator.vcxproj:
* WebKit.vcxproj/WebKitExportGenerator/WebKitExportGenerator.vcxproj.filters:
Make WebKitExports.def.in be treated as a custom build file so that changes to it cause the exports to be rebuilt.

* JavaScriptCore.vcxproj/JavaScriptCoreExportGenerator/JavaScriptCoreExportGenerator.vcxproj:
* JavaScriptCore.vcxproj/JavaScriptCoreExportGenerator/JavaScriptCoreExportGenerator.vcxproj.filters:
Make JavaScriptCoreExports.def.in be treated as a custom build file so that changes to it cause the exports to be rebuilt.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@142203 268f45cc-cd09-0410-ab3c-d52691b4dbfc